requirements.txt
streamlit
google.generativeai
st_pages